Fix assertion.
git-svn-id: https://svn.dealii.org/trunk@24002 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/include/deal.II/base/tensor.h
@@ -519,7 +519,9 @@ inline
 void
 Tensor<rank_, dim, Number>::unroll (Vector<Number2> &result) const
 {
-  AssertDimension (result.size(),std::pow(static_cast<double>(dim),rank_));
+  AssertDimension (result.size(),
+		   static_cast<unsigned int>(std::pow(static_cast<double>(dim),
+						      rank_)));
   unsigned index = 0;
   unroll_recursion (result, index);
 }
